quinn: move UdpExt functionality into platform-specific UdpSocket types
