+++ quinn/src/platform/mod.rs
@@ -1,8 +1,10 @@
 //! Uniform interface to send/recv UDP packets with ECN information.
-use proto::Transmit;
-use std::{io, io::IoSliceMut};
+use std::{
+    io::{self, IoSliceMut},
+    net::{IpAddr, Ipv6Addr, SocketAddr},
+};
 
-use crate::udp::RecvMeta;
+use proto::{EcnCodepoint, Transmit};
 
 #[cfg(unix)]
 mod cmsg;
@@ -14,6 +16,8 @@ mod imp;
 #[cfg(not(unix))]
 #[path = "fallback.rs"]
 mod imp;
+
+pub use imp::UdpSocket;
 
 #[allow(dead_code)] // TODO: Remove when used
 /// Returns the platforms UDP socket capabilities
@@ -35,4 +39,25 @@ pub trait UdpExt {
 pub struct UdpCapabilities {
     /// Whether the platform supports Generic Send Offload (GSO)
     pub gso: bool,
+}
+
+#[derive(Debug, Copy, Clone)]
+pub struct RecvMeta {
+    pub addr: SocketAddr,
+    pub len: usize,
+    pub ecn: Option<EcnCodepoint>,
+    /// The destination IP address which was encoded in this datagram
+    pub dst_ip: Option<IpAddr>,
+}
+
+impl Default for RecvMeta {
+    /// Constructs a value with arbitrary fields, intended to be overwritten
+    fn default() -> Self {
+        Self {
+            addr: SocketAddr::new(Ipv6Addr::UNSPECIFIED.into(), 0),
+            len: 0,
+            ecn: None,
+            dst_ip: None,
+        }
+    }
 }
